close #920
Media file urls should not use the static tag, that is only for static
assets to get their static url.  When the files are stored locally, the
media url is its path, so this happens to work, but when using S3,
trying to convert the media url to a static url fails.

Media files = user uploaded
Static files = check in to git