fixes navigation rendering on mobile
This fixes a bug in iOS browsers where the bottom navigation would not
stay anchored to the bottom of the viewport; instead, it would snap and
pop around when the viewport size was changed. This was fixed by
independently fixing the top and bottom navigation items to the viewport
when the bottom navigation is visible at smaller sizes, then fixing the entire navigation
element when it is unified at larger sizes.