Document required step
Although most of the Invoke tool's tasks are defined in such a way to
explicitly specify the `local.yml` file via the `-f` flag of the
`docker-compose` utility, the task named "up" is not:

    @task
    def up(c):
        """Start the docker images"""
        c.run("docker-compose up -d")

Because of this, the command `inv up` will not function as intended on
its own.

This is similar to how other MuckRock projects are organized as of this
writing. For consistency with those projects, instruct contributors to
set the `COMPOSE_FILE` environment variable in all cases.