Correct and contextualize command
Even when run in a directory with an executable file named `manage.py`,
the following command will fail in Bash:

    manage.py creatersakey

...because the current working directory is not included in the `PATH`
environment variable. Replace with the Squarelet project's "task" named
"manage", using the "createrskkey" parameter.

Expand the instructions to explicitly explain how to execute a command
"from the Squarelet command line."